[Fix] NewsFeed query & mutation

+++ src/schema/resolvers/newsFeed/newsCommentResolvers.js
@@ -13,8 +13,11 @@ export default {
     getAllNewsComments: async (
       root,
       data,
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const newsComment = await NewsComment.query();
         return newsComment;
@@ -27,8 +30,11 @@ export default {
     getNewsCommentByID: async (
       root,
       { id },
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const newsComment = await NewsComment.query().findById(id);
         if (!newsComment) {
@@ -38,17 +44,17 @@ export default {
       } catch (e) {
         // eslint-disable-next-line no-console
         console.error(e);
-        if (e.message === 'newsComment-not-found') {
-          throw new ValidationError('newsComment-not-found');
-        }
         throw new ValidationError('bad-request');
       }
     },
     getNewsCommentByNewsID: async (
       root,
       { news_id },
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const newsComment = await NewsComment.query().where('news_id', news_id);
         if (!newsComment) {
@@ -58,17 +64,17 @@ export default {
       } catch (e) {
         // eslint-disable-next-line no-console
         console.error(e);
-        if (e.message === 'newsComment-not-found') {
-          throw new ValidationError('newsComment-not-found');
-        }
-        throw new ValidationError(e);
+        throw new ValidationError('bad-request');
       }
     },
     getNewsCommentByUserID: async (
       root,
       { user_id },
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const newsComment = await NewsComment.query().where('user_id', user_id);
         if (!newsComment) {
@@ -78,9 +84,6 @@ export default {
       } catch (e) {
         // eslint-disable-next-line no-console
         console.error(e);
-        if (e.message === 'newsComment-not-found') {
-          throw new ValidationError('newsComment-not-found');
-        }
         throw new ValidationError('bad-request');
       }
     },
@@ -89,8 +92,11 @@ export default {
     insertNewsComment: async (
       root,
       data,
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const newsCommentInsert = await NewsComment.query().insert(data);
         return newsCommentInsert;
@@ -103,8 +109,11 @@ export default {
     updateNewsComment: async (
       root,
       data,
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const updateNewsComment = await NewsComment.query().updateAndFetchById(
           data.id,
@@ -120,8 +129,11 @@ export default {
     deleteNewsComment: async (
       root,
       { id },
-      { models: { NewsComment }, ValidationError },
+      { models: { NewsComment }, ValidationError, user },
     ) => {
+      if (!user) {
+        throw new ValidationError('unauthorized');
+      }
       try {
         const deleteNewsComment = await NewsComment.query().findById(id);
         await NewsComment.query().deleteById(id);
